requirements.txt
asgiref==3.5.2
dj-config-url==0.1.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
env==0.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
